Refactor for command modules
Co-local code for a given example.
